+++ Assets/Scripts/Generation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: de207d6bae62005409a2cd030a5641c9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
