+++ Assets/Sprites/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9867741e1b5f36744af6c2210a818637
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
