+++ Assets/Sprites/UI/trophy.png.meta
