--- Assets/TitleToMain.cs
@@ -1,23 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-//Purpose: literally just to change from title screen to menu screen
-//Usage: put this on an empty game object and then drag it into the scene 
-public class TitleToMain : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.Space))
-        {
-            SceneManager.LoadScene("MainMenu");
-        }
-    }
-}
